requirements.txt
flask-restplus==0.13.0
werkzeug==0.16.1
flask-cors==3.0.8
Flask-PyMongo==2.3.0
mongodb-migrations==0.7.0
